+++ Makefile
@@ -7,16 +7,13 @@ INCLUDES = -Iinclude -Ilib/HT16K33Disp -Ilib/Randomizer -Inative
 
 # Source files
 SOURCES = src/buffers.cpp \
-          src/chime.cpp \
           src/contrast.cpp \
-          src/contrast_handler.cpp \
+          src/contrast_handler.cpp\
           src/displays.cpp \
           src/event_dispatcher.cpp \
           src/leds.cpp \
-          src/led_handler.cpp \
-          src/mode.cpp \
+          src/led_handler.cpp \          src/mode.cpp \
           src/mode_handler.cpp \
-          src/morse.cpp \
           src/option.cpp \
           src/options_presenter.cpp \
           src/option_handler.cpp \
