+++ deprecated/chime.h
