+++ deprecated/morse.h
